api: Make create_book a module
