+++ src/api/src/create_book/generators/__init__.py
@@ -1,0 +1,2 @@
+from epub import EPUBGenerator
+from pdf import PDFGenerator
